fix various random warnings

+++ NorthstarDedicatedTest/modmanager.cpp
@@ -333,7 +333,7 @@ void ModManager::LoadMods()
 	}
 
 	// in a seperate loop because we register mod files in reverse order, since mods loaded later should have their files prioritised
-	for (int i = m_loadedMods.size() - 1; i > -1; i--)
+	for (size_t i = m_loadedMods.size() - 1; i > -1; i--)
 	{
 		if (!m_loadedMods[i].Enabled)
 			continue;
